Add createVariant mutation and separate types into files to deal with circular dep
