+++ site/gatsby-site/server/types/entity.ts
@@ -1,0 +1,13 @@
+import { GraphQLNonNull, GraphQLObjectType, GraphQLString } from "graphql";
+import { ObjectIdScalar } from "../scalars";
+import { GraphQLDateTime } from "graphql-scalars";
+
+export const EntityType = new GraphQLObjectType({
+    name: 'Entity',
+    fields: {
+        _id: { type: ObjectIdScalar },
+        entity_id: { type: new GraphQLNonNull(GraphQLString) },
+        name: { type: new GraphQLNonNull(GraphQLString) },
+        created_at: { type: GraphQLDateTime },
+    },
+});
